refactor: move experiences to resume component
