+++ src/data/index.ts
@@ -1,2 +1,1 @@
-export * from './experiences'
 export * from './routes'
